+++ src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/controller/model/DashboardVisualisationColumnsDefinition.kt
@@ -1,0 +1,8 @@
+package uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.model
+
+data class DashboardVisualisationColumnsDefinition(
+  val keys: List<DashboardVisualisationColumnDefinition>? = null,
+  val measures: List<DashboardVisualisationColumnDefinition>,
+  val filters: List<ValueVisualisationColumnDefinition>? = null,
+  val expectNulls: Boolean,
+)