+++ src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/controller/model/DashboardVisualisationDefinition.kt
@@ -1,0 +1,9 @@
+package uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.model
+
+data class DashboardVisualisationDefinition(
+  val id: String,
+  val type: DashboardVisualisationTypeDefinition,
+  val display: String? = null,
+  val description: String? = null,
+  val columns: DashboardVisualisationColumnsDefinition,
+)