+++ src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/data/model/DashboardVisualisation.kt
@@ -1,0 +1,9 @@
+package uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model
+
+data class DashboardVisualisation(
+  val id: String,
+  val type: DashboardVisualisationType,
+  val display: String? = null,
+  val description: String? = null,
+  val columns: DashboardVisualisationColumns,
+)